Multiple views exposed filter handling bugfix.
When multiple views exposed filters, in this specific bug case exo exposed filters, this.$exposed_form context would return undefined.

This would cause only one exposed filter to be extended, and ajax would fail on others within the same page.

For example an exposed filter, and that same exposed filter higher on the page.

This does resolve the issue, and perhaps is the best solution, as an exposed filter could be anywhere within the html body.